# Conflicts:
#	backend/src/main/kotlin/ch/zhaw/pm4/simonsays/api/controller/MenuItemController.kt
#	backend/src/main/kotlin/ch/zhaw/pm4/simonsays/api/mapper/IngredientMapper.kt
#	backend/src/main/kotlin/ch/zhaw/pm4/simonsays/api/mapper/MenuItemMapper.kt
#	backend/src/main/kotlin/ch/zhaw/pm4/simonsays/api/types/MenuItemCreateUpdateDTO.kt
#	backend/src/main/kotlin/ch/zhaw/pm4/simonsays/api/types/MenuItemDTO.kt
#	backend/src/main/kotlin/ch/zhaw/pm4/simonsays/entity/Event.kt
#	backend/src/main/kotlin/ch/zhaw/pm4/simonsays/entity/MenuItem.kt
#	backend/src/main/kotlin/ch/zhaw/pm4/simonsays/repository/IngredientRepository.kt
#	backend/src/main/kotlin/ch/zhaw/pm4/simonsays/repository/MenuItemRepository.kt
#	backend/src/test/kotlin/ch/zhaw/pm4/simonsays/MenuItemIntegrationTest.kt
#	backend/src/test/kotlin/ch/zhaw/pm4/simonsays/MenuItemTest.kt
#	backend/src/test/kotlin/ch/zhaw/pm4/simonsays/MenuItemTestData.kt
#	backend/src/test/kotlin/ch/zhaw/pm4/simonsays/factory/MenuItemFactory.kt

+++ backend/src/main/kotlin/ch/zhaw/pm4/simonsays/entity/MenuItem.kt
@@ -20,7 +20,7 @@ data class MenuItem (
         var ingredients: List<Ingredient>,
 
         @Column
-        var price: Long,
+        var price: Double,
 
         @ManyToMany(mappedBy = "menuItems")
         val menus: List<Menu>?
